+++ astra_config/__init__.py
@@ -1,0 +1,3 @@
+﻿from astra_config.settings import AstraSettings
+
+__all__ = ["AstraSettings"]